Add usize matrix type
TSPLIB distances are rounded to the nearest integer and ought to
be stored as such. `value` in AntResult was changed accordingly.

+++ src/aco/result_log.rs
@@ -17,7 +17,7 @@ impl TimestampedResult {
         }
     }
 
-    fn value(&self) -> f64 {
+    fn value(&self) -> usize {
         self.result.value
     }
 }
@@ -35,7 +35,7 @@ impl ResultLog {
         }
     }
 
-    pub fn best_value(&self) -> f64 {
+    pub fn best_value(&self) -> usize {
         self.log[self.best_so_far].value()
     }
